Relative Tracing Offset Enables Arbitrarily Massive Worlds
I drove out to +Y 22763101.76675415 in a few seconds, and everything
looked fine.  No noticable seams driving at low speeds either.  This
is a true infinitely repeating world.

If my calculations are correct, the coordinate system supports
distances up to 1881.5 astronomical units, which a constant precision
of approximately 15 micrometers.  This comes with no significant
impact on frame rate or throughput.